change: rename core_native_gba.h to zeebo.h
