+++ src/draw_line.c
@@ -1,4 +1,4 @@
-#include "core_native_gba.h"
+#include "zeebo.h"
 
 void draw_cmd_line(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2)
 {
